backend/requirements.txt
connexion[flask,uvicorn,swagger-ui] == 3.1.0
gunicorn==23.0.0
pymongo==4.10.0
python-dotenv==1.0.1
six==1.16.0
pytz==2024.2
#setuptools >= 75.3.0
#Flask-Cors==5.0.0
Flask-APScheduler==1.13.1
Flask-Mail==0.10.0

